+++ packages/emotion/src/index.js
@@ -1,4 +1,5 @@
 import { hashString, Stylis, memoize, unitless } from 'emotion-utils'
+import stylisPluginEmotion from 'stylis-plugin-emotion'
 import StyleSheet from './sheet'
 
 export const sheet = new StyleSheet()
@@ -16,6 +17,12 @@ const externalStylisPlugins = []
 
 const use = stylis.use
 
+function insertRule(rule) {
+  sheet.insert(rule, currentSourceMap)
+}
+
+const insertionPlugin = stylisPluginEmotion(insertRule)
+
 export const useStylisPlugin = plugin => {
   externalStylisPlugins.push(plugin)
   use(null)(externalStylisPlugins)(insertionPlugin)
@@ -25,96 +32,7 @@ export let registered = {}
 
 export let inserted = {}
 
-export function flush() {
-  sheet.flush()
-  inserted = {}
-  registered = {}
-  sheet.inject()
-}
-
-function insertRule(rule) {
-  sheet.insert(rule, currentSourceMap)
-}
-
 let currentSourceMap = ''
-let queue = []
-let parentQueue = []
-
-function insertionPlugin(
-  context,
-  content,
-  selectors,
-  parents,
-  line,
-  column,
-  length,
-  id
-) {
-  switch (context) {
-    case -2: {
-      queue.forEach(insertRule)
-      queue = []
-      parentQueue = []
-      break
-    }
-
-    case 2: {
-      if (id === 0) {
-        const selector = selectors.join(',')
-        let parent = parents.join(',')
-        const rule = `${selector}{${content}}`
-        let index = parentQueue.indexOf(selector)
-        if (index === -1) {
-          index = parentQueue.length
-        } else {
-          let length = queue.length
-          while (length--) {
-            if (parentQueue[length] === selector) {
-              parentQueue[length] = undefined
-            }
-          }
-        }
-        queue.splice(index, 0, rule)
-        parentQueue.splice(index, 0, parent)
-      }
-      break
-    }
-    // after an at rule block
-    case 3: {
-      let parent = parents.join(',')
-      parentQueue.push(parent)
-      let chars = selectors.join('')
-      const second = chars.charCodeAt(1)
-      let child = content
-      switch (second) {
-        // s upports
-        case 115:
-        // d ocument
-        // eslint-disable-next-line no-fallthrough
-        case 100:
-        // m edia
-        // eslint-disable-next-line no-fallthrough
-        case 109: {
-          queue.push(chars + '{' + child + '}')
-          break
-        }
-        // k eyframes
-        case 107: {
-          chars = chars.substring(1)
-          child = chars + '{' + child + '}'
-          queue.push('@-webkit-' + child)
-          queue.push('@' + child)
-          parentQueue.push(parent)
-          break
-        }
-        default: {
-          queue.push(chars + child)
-          break
-        }
-      }
-    }
-  }
-}
 
 stylis.use(insertionPlugin)
 
@@ -324,4 +242,11 @@ export function hydrate(ids) {
   ids.forEach(id => {
     inserted[id] = true
   })
+}
+
+export function flush () {
+  sheet.flush()
+  inserted = {}
+  registered = {}
+  sheet.inject()
 }
